Port over installation instructions
There are a few caveats:

I used @VERSION@ all the places where 1.5.1 was mentioned on the site.
On my branch that's currently 1.6.0-RC1-14-b09fd59a. I don't know
if we need to distinguish between the current version and the stable
version

The --help section is a bit funny as it's shown as a REPL session now
and therefore some of the text is truncated. I haven't used tut before,
is there a better way to do this?

I haven't ported the intellij section over as I believe it's now built
in?